fix(livestore): add missing store option
